+++ conv.go
@@ -5,6 +5,7 @@ const (
 	maxUint32 = uint64(^uint32(0))
 )
 
+// i32FromInt converts an int to an int32.
 func i32FromInt(n int) (int32, error) {
 	if n < 0 || n > maxInt32 {
 		return 0, ErrSizeOverflow
@@ -13,6 +14,7 @@ func i32FromInt(n int) (int32, error) {
 	return int32(n), nil
 }
 
+// u32FromInt converts an int to a uint32.
 func u32FromInt(n int) (uint32, error) {
 	if n < 0 || uint64(n) > maxUint32 {
 		return 0, ErrSizeOverflow